chore: prepare QueryApi async
